requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
